feat: add travel type
