+++ src/main/java/com/ajou/travely/domain/travel/TravelType.java
@@ -1,0 +1,11 @@
+package com.ajou.travely.domain.travel;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum TravelType {
+    PUBLIC,
+    PRIVATE
+}
