+++ src/main/java/com/ajou/travely/service/CostService.java
@@ -4,7 +4,7 @@ import com.ajou.travely.controller.cost.dto.CostCreateResponseDto;
 import com.ajou.travely.controller.cost.dto.CostResponseDto;
 import com.ajou.travely.controller.cost.dto.*;
 import com.ajou.travely.domain.Cost;
-import com.ajou.travely.domain.Travel;
+import com.ajou.travely.domain.travel.Travel;
 import com.ajou.travely.domain.UserCost;
 import com.ajou.travely.exception.ErrorCode;
 import com.ajou.travely.domain.user.User;
