+++ src/test/java/com/ajou/travely/service/CostServiceTest.java
@@ -7,7 +7,7 @@ import com.ajou.travely.controller.cost.dto.CostUpdateInfoDto;
 import com.ajou.travely.domain.Cost;
 import com.ajou.travely.controller.cost.dto.UserCostResponseDto;
 import com.ajou.travely.controller.travel.dto.TravelCreateRequestDto;
-import com.ajou.travely.domain.Travel;
+import com.ajou.travely.domain.travel.Travel;
 import com.ajou.travely.domain.UserCost;
 import com.ajou.travely.domain.user.Type;
 import com.ajou.travely.domain.user.User;
@@ -16,16 +16,11 @@ import com.ajou.travely.repository.TravelRepository;
 import com.ajou.travely.repository.UserCostRepository;
 import com.ajou.travely.repository.UserRepository;
 import org.assertj.core.api.Assertions;
-import org.junit.Assert;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
