+++ src/test/java/com/ajou/travely/service/PostServiceTest.java
@@ -7,7 +7,7 @@ package com.ajou.travely.service;
 //import com.ajou.travely.domain.Place;
 //import com.ajou.travely.domain.Post;
 //import com.ajou.travely.domain.Schedule;
-//import com.ajou.travely.domain.Travel;
+//import com.ajou.travely.domain.travel.Travel;
 //import com.ajou.travely.domain.user.Type;
 //import com.ajou.travely.domain.user.User;
 //import com.ajou.travely.repository.PlaceRepository;
